+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Note that it's not meant to be run as a Python library, so there are no guarante
 
 - add support for [PEP-735](https://peps.python.org/pep-0735/) dependency groups
 - add support for finding `pytest` in uv's default dependency group ([docs](https://docs.astral.sh/uv/concepts/projects/dependencies/#dependency-groups))
+- run `pytest` via popular package managers (`uv`, `poetry`, `pdm`) if a matching lockfile is found (e.g. `uv run pytest` is favored over `pytest` if `uv.lock` is present)
 
 ## 0.6.2